+++ Source/SupertalkEditor/SupertalkEditorSettings.cpp
@@ -1,0 +1,9 @@
+﻿// Copyright (c) MissiveArts LLC
+
+#include "SupertalkEditorSettings.h"
+
+USupertalkEditorSettings::USupertalkEditorSettings()
+{
+	bEnableScriptEditor = false;
+	bSaveSourceFilesInScriptEditor = false;
+}